+++ .gitignore
@@ -6,3 +6,5 @@
 .vimrc
 
 Backstage/test/test_quick.js
+
+*/dist/
