+++ Backstage/.eslintrc
@@ -7,6 +7,9 @@
    "extends": [
       "eslint:recommended",
       "airbnb-base",
+      "plugin:import/errors",
+      "plugin:import/warnings",
+      "plugin:import/typescript",
       "plugin:@typescript-eslint/eslint-recommended",
       "plugin:@typescript-eslint/recommended"
    ],
@@ -47,7 +50,10 @@
       "import/resolver": {
          "typescript": {
             "alwaysTryTypes": true,
-            "project": "/"
+            "project": "./tsconfig.json"
+         },
+         "node": {
+            "extensions": [".ts"]
          }
       }
    }
