+++ Backstage/package.json
@@ -12,6 +12,7 @@
    "author": "",
    "license": "ISC",
    "devDependencies": {
+      "@types/node": "^18.6.3",
       "@typescript-eslint/eslint-plugin": "^5.32.0",
       "@typescript-eslint/parser": "^5.32.0",
       "eslint": "^8.21.0",
