+++ Backstage/src/executor/game_context.ts
@@ -134,7 +134,3 @@ const buildGameContext = ruleSet => ({
    computedSkills: null,
    computedAscensionPerks: null
 })
-
-module.exports = {
-   buildGameContext
-}
